Implement validation

+++ AutoRest/Generators/Java/Java.Tests/src/test/java/fixtures/bodycomplex/BasicOperationsTests.java
@@ -20,7 +20,7 @@ public class BasicOperationsTests {
     @Test
     public void getValid() throws Exception {
         Basic result = client.getBasicOperations().getValid();
-        Assert.assertEquals(2, result.getId());
+        Assert.assertEquals(2, result.getId().intValue());
         Assert.assertEquals("abc", result.getName());
         Assert.assertEquals(CMYKColors.YELLOW, result.getColor());
     }
