+++ AutoRest/Generators/Java/Java.Tests/src/test/java/fixtures/bodycomplex/InheritanceTests.java
@@ -24,7 +24,7 @@ public class InheritanceTests {
         Siamese result = client.getInheritance().getValid();
         Assert.assertEquals("persian", result.getBreed());
         Assert.assertEquals("green", result.getColor());
-        Assert.assertEquals(2, result.getId());
+        Assert.assertEquals(2, result.getId().intValue());
         Assert.assertEquals("Siameeee", result.getName());
         Assert.assertEquals("french fries", result.getHates().get(1).getFood());
     }
